Fix PTR2INT/PTR2UINT, so they no longer are restricted to the "int" range. Make TCL_Z_MODIFIER available in Tk (even when compiled with Tcl 8.6), and use it. More size_t/clientData related improvements

+++ macosx/tkMacOSXBitmap.c
@@ -173,7 +173,7 @@ TkpCreateNativeBitmap(
     OSErr err;
 
     err = ChkErr(GetIconRef, kOnSystemDisk, kSystemIconsCreator,
-	    PTR2UINT(source), &icon);
+	    (unsigned int)PTR2UINT(source), &icon);
     if (err == noErr) {
 	pixmap = GetBitmapForIcon(display, icon, CGSizeMake(builtInIconSize,
 		builtInIconSize));
